product Deep link fixing

+++ el-grocer-shopper-sdk-iOS/Classes/Sources/Model/Entities+CRUD/Product+CRUD.swift
@@ -521,15 +521,18 @@ extension Product {
                     }else{
                         product.promotionOnly = false
                         // deleting because product is not available in shops dictionary for current store https://elgrocerdxb.atlassian.net/browse/EEN-4110
-                        context.delete(product)
-                        continue
+                        if onlyCurrentStoreProducts {
+                            context.delete(product)
+                            continue
+                        }
+                       
                     }
                     
                 }else{
                     product.promotionOnly = false
                 }
 
-                if product.promotionOnly.boolValue && product.promotion == 0 {
+                if product.promotionOnly.boolValue && product.promotion == 0 && onlyCurrentStoreProducts {
                     context.delete(product)
                     continue
                 }
